refactor: embed espanso-gui as a library into espanso binary
- espanso-gui/src/main.rs → lib.rs: main() becomes run()
- espanso/Cargo.toml: add espanso-gui as dependency
- espanso/src/cli/gui.rs: spawn GUI on dedicated thread via
  espanso_gui::run() instead of spawning external binary
- Only one binary now: espanso

espanso gui → runs GUI embedded, no separate espanso-gui binary
