+++ espanso/src/cli/gui.rs
@@ -17,8 +17,9 @@
  * along with espanso.  If not, see <https://www.gnu.org/licenses/>.
  */
 
+use std::thread;
+
 use log::info;
-use std::process::Command;
 
 use super::{CliModule, CliModuleArgs};
 
@@ -36,63 +37,27 @@ pub fn new() -> CliModule {
 fn gui_main(args: CliModuleArgs) -> i32 {
     let paths = args.paths.expect("missing paths in gui main");
 
-    // Determine the path to the espanso-gui binary
-    // It should be in the same directory as the main espanso binary
-    let gui_binary = find_gui_binary();
-
     let initial_module = args
         .cli_args
         .as_ref()
         .and_then(|m| m.value_of("module"))
-        .unwrap_or("match");
+        .map(|m| espanso_gui::app::Module::from_str(m))
+        .unwrap_or(espanso_gui::app::Module::MatchManager);
+
+    let config_dir = paths.config.clone();
+    let runtime_dir = paths.runtime.clone();
 
     info!(
-        "launching espanso management GUI: {} --config_dir={} --runtime_dir={} {}",
-        gui_binary.display(),
-        paths.config.display(),
-        paths.runtime.display(),
-        initial_module
+        "launching espanso management GUI (embedded), config={}",
+        config_dir.display()
     );
 
-    match Command::new(&gui_binary)
-        .arg(format!("--config_dir={}", paths.config.display()))
-        .arg(format!("--runtime_dir={}", paths.runtime.display()))
-        .arg(initial_module)
-        .spawn()
-    {
-        Ok(_child) => {
-            info!("GUI process spawned successfully");
-            // Don't wait — let the GUI run independently
-            0
-        }
-        Err(e) => {
-            // If the GUI binary isn't found, print a helpful message
-            eprintln!("Unable to launch espanso GUI: {e}");
-            eprintln!(
-                "The espanso-gui binary was not found at: {}",
-                gui_binary.display()
-            );
-            eprintln!("Make sure espanso-gui is installed alongside the main espanso binary.");
-            1
-        }
-    }
-}
+    // Run the GUI on a dedicated thread so it doesn't block the CLI.
+    // The egui event loop will drive its own run loop.
+    thread::spawn(move || {
+        espanso_gui::run(Some(config_dir), Some(runtime_dir), initial_module);
+    });
 
-/// Find the espanso-gui binary.
-///
-/// On all platforms, it's expected to be in the same directory as the
-/// main espanso executable.
-fn find_gui_binary() -> std::path::PathBuf {
-    let current_exe =
-        std::env::current_exe().expect("unable to determine current executable path");
-    let exe_dir = current_exe
-        .parent()
-        .expect("unable to determine executable directory");
-
-    #[cfg(target_os = "windows")]
-    let gui_name = "espanso-gui.exe";
-    #[cfg(not(target_os = "windows"))]
-    let gui_name = "espanso-gui";
-
-    exe_dir.join(gui_name)
+    // Return immediately — the GUI runs in the background
+    0
 }
